Moved logic to add new component
